Add mono-Hydrogen (atomic hydrogen) resource hdf5
The most simple hdf5 file possible. 1 spatial orbital, 2 spin-orbitals, 1 electron filling the orbital. The exact groundstate is degenerate with any linear combination of electron in either spin-orbital having the same energy. The hf, ccsd and fci energies are all exact and equal to -0.4665818495572751 Hartree.
Simulation should be done on 2 qubits. The Hartree-Fock (HF) state is then produced by an X gate on one qubit and nothing on the other qubit.
Note that the UCCSD operator would be empty. The code should probably produce an empty circuit then, for sake of completeness, or warn that the user should select HF method instead.